[MDH-105] refactor : 상점 생성시 전화번호, 위도, 경도, 정규식 추가
