+++ src/main/java/com/mdh/devtable/global/util/RegularExpression.java
@@ -8,4 +8,9 @@ public class RegularExpression {
 
     public static final String PHONE_NUMBER = "^01([0|1|6|7|8|9]{1})[0-9]{3,4}[0-9]{4}$";
 
+    public static final String SHOP_TEL_NUMBER = "^[0-9]{10,11}$";
+    
+    public static final String URL = "^(http|https)://.*";
+
+
 }
